Switch dependency to browsergym-core (#5242)
