+++ pyproject.toml
@@ -28,7 +28,7 @@ uvicorn = "*"
 types-toml = "*"
 numpy = "*"
 json-repair = "*"
-browsergym = "0.10.2" # integrate browsergym as the browsing interface
+browsergym-core = "0.10.2" # integrate browsergym-core as the browsing interface
 html2text = "*"
 e2b = ">=0.17.1,<1.1.0"
 pexpect = "*"
@@ -63,6 +63,7 @@ opentelemetry-exporter-otlp-proto-grpc = "1.25.0"
 modal = "^0.66.26"
 runloop-api-client = "0.10.0"
 pygithub = "^2.5.0"
+joblib = "*"
 openhands-aci = "^0.1.1"
 python-socketio = "^5.11.4"
 redis = "^5.2.0"
@@ -142,6 +143,8 @@ gdown = "*"
 matplotlib = "*"
 seaborn = "*"
 tabulate = "*"
+browsergym-webarena = "0.10.2"
+browsergym-miniwob = "0.10.2"
 
 [tool.poetry-dynamic-versioning]
 enable = true
